bump to 1.0.0
This library is stable enough and the recent issues regarding changing to newer rubies
have proven enough people are using it to warrant following semantic version conventions.

As per Semantic versioning though all bets are off in a pre 1.0.0 release.
So this is to make an official commitment to avoiding backwards incompatible changes.

Any dropping of currently supported rubies will be handled more carefully.

It also looks like this is now over 10 years old, so it is perhaps not too soon to make
this change!